[haproxy] Support specification of members via tags
Add a new tag haproxy::balancemember that is used to indicate to
haproxy that this node should be a member of the listen stanza for
the relevant service. The omission of this tag will result in the
exclusion of a node, which allows haproxy itself to register that
it is providing a service and not attempt to loadbalance itself.

Add a new tag haproxy::balanced that indicates that a particular
service is behind a load balancer when accessed at the registered
address. This will not be very useful until multiple address
advertisement is made available to consul upstream.